+++ src/Symfony/Component/Config/CHANGELOG.md
@@ -1,12 +1,16 @@
 CHANGELOG
 =========
 
+4.3.0
+-----
+
+ * deprecated using environment variables with `cannotBeEmpty()` if the value is validated with `validate()`
+
 4.2.0
 -----
 
  * deprecated constructing a `TreeBuilder` without passing root node information
  * renamed `FileLoaderLoadException` to `LoaderLoadException`
- * deprecated using environment variables with `cannotBeEmpty()` if the value is validated with `validate()`
 
 4.1.0
 -----